# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MeritView API
  version: 1.0.0
  description: AI Decision Support Platform API
  contact:
    name: MeritView API Support
    url: https://meritview.app/support
    email: api@meritview.app
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/v1
    description: Development server
  - url: https://api.staging.meritview.app/v1
    description: Staging server
  - url: https://api.meritview.app/v1
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DisputeCategory:
      type: string
      enum: [contract_interpretation, small_claims_assessment, partnership_conflict]
      example: contract_interpretation
    DisputeState:
      type: string
      enum: [draft, brief_submitted, payment_pending, under_analysis, awaiting_aggregation, completed, withdrawn, declined, failed, awaiting_counterparty, in_progress, awaiting_briefs, awaiting_counterparty_brief]
      example: draft
    PricingTier:
      type: string
      enum: [standard, expedited, extended, reanalysis]
      example: standard
    PartyRole:
      type: string
      enum: [initiator, respondent]
      example: initiator
...
    Dispute:
      type: object
      properties:
        id:
          type: string
          pattern: ^disp_[a-z0-9]{6}$
          example: disp_abc123
        category:
          $ref: '#/components/schemas/DisputeCategory'
        title:
          type: string
          maxLength: 200
          example: Client refusing to pay $8,500 consulting invoice
        summary:
          type: string
          maxLength: 500
          nullable: true
          example: Client claims scope was exceeded; contract language is ambiguous
        estimatedStakesUsd:
          type: number
          format: decimal
          nullable: true
          example: 8500
        state:
          $ref: '#/components/schemas/DisputeState'
        pricingTier:
          $ref: '#/components/schemas/PricingTier'
        priceUsd:
          type: number
          format: decimal
          example: 99.00
        parties:
          type: array
          items:
            $ref: '#/components/schemas/Party'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-17T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-17T10:30:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - category
        - title
        - state
        - pricingTier
        - priceUsd
        - parties
        - createdAt
        - updatedAt
    CreateDisputeRequest:
      type: object
      required:
        - category
        - title
      properties:
        category:
          $ref: '#/components/schemas/DisputeCategory'
        title:
          type: string
          minLength: 5
          maxLength: 200
        summary:
          type: string
          maxLength: 500
        estimatedStakesUsd:
          type: number
          format: decimal
          minimum: 0
        counterparty:
          type: object
          properties:
            email:
              type: string
              format: email
            displayNameForInvitation:
              type: string
              maxLength: 100
          required:
            - email
        pricingTier:
          $ref: '#/components/schemas/PricingTier'
          default: standard
    CreateDisputeResponse:
      type: object
      properties:
        dispute:
          $ref: '#/components/schemas/Dispute'
        paymentIntent:
          type: object
          properties:
            clientSecret:
              type: string
              example: pi_xyz_secret_abc...
          required:
            - clientSecret
    DisputeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dispute'
        pagination:
          type: object
          properties:
            nextCursor:
              type: string
              nullable: true
            hasMore:
              type: boolean
          required:
            - nextCursor
            - hasMore
    DisputeDetailResponse:
      type: object
      properties:
        dispute:
          allOf:
            - $ref: '#/components/schemas/Dispute'
            - type: object
              properties:
                briefs:
                  type: array
                  items:
                    $ref: '#/components/schemas/Brief'
                opinion:
                  $ref: '#/components/schemas/Opinion'
                evaluatorOutputs:
                  type: array
                  items:
                    $ref: '#/components/schemas/EvaluatorOutput'
                payments:
                  type: array
                  items:
                    $ref: '#/components/schemas/Payment'
    UpdateDisputeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        summary:
          type: string
          maxLength: 500
        estimatedStakesUsd:
          type: number
          format: decimal
          minimum: 0
    WithdrawResponse:
      type: object
      properties:
        dispute:
          $ref: '#/components/schemas/Dispute'
...
    Brief:
      type: object
      properties:
        id:
          type: string
          pattern: ^brief_[a-z0-9]{3}$
          example: brief_abc
        partyId:
          type: string
          pattern: ^party_[a-z0-9]{3}$
        sections:
          type: object
          properties:
            factual_background:
              type: string
            my_position:
              type: string
            supporting_arguments:
              type: string
            acknowledgment_of_opposing:
              type: string
            desired_resolution:
              type: string
          required:
            - factual_background
            - my_position
            - supporting_arguments
            - acknowledgment_of_opposing
            - desired_resolution
        wordCount:
          type: integer
          example: 1247
        supportingDocumentIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, submitted, sealed]
        createdAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
          nullable: true
        sealedAt:
          type: string
          format: date-time
          nullable: true
...
    Opinion:
      type: object
      properties:
        id:
          type: string
          pattern: ^op_[a-z0-9]{3}$
          example: op_xyz
        disputeId:
          type: string
        generatedAt:
          type: string
          format: date-time
        promptVersion:
          type: string
          example: eval-v3.2, agg-v2.1
        evaluatorsUsed:
          type: array
          items:
            type: string
            example: anthropic/claude-3-5-sonnet
        executiveSummary:
          type: string
        keyIssues:
          type: array
          items:
            type: object
            properties:
              issue:
                type: string
              agreementLevel:
                type: string
                enum: [high, medium, low]
        partyAAnalysis:
          type: object
          properties:
            strongestArguments:
              type: array
              items:
                type: string
            weakestPoints:
              type: array
              items:
                type: string
            factualConcerns:
              type: array
              items:
                type: string
        partyBAnalysis:
          type: object
          properties:
            strongestArguments:
              type: array
              items:
                type: string
            weakestPoints:
              type: array
              items:
                type: string
            factualConcerns:
              type: array
              items:
                type: string
        comparativeAssessment:
          type: string
        confidenceIndicators:
          type: object
          properties:
            overallConfidence:
              type: number
              format: float
              minimum: 0
              maximum: 1
            evaluatorAgreement:
              type: number
              format: float
              minimum: 0
              maximum: 1
          required:
            - overallConfidence
            - evaluatorAgreement
        suggestedConsiderations:
          type: object
          properties:
            partyA:
              type: array
              items:
                type: string
            partyB:
              type: array
              items:
                type: string
        disclaimers:
          type: array
          items:
            type: string
          minItems: 4
          example:
            - "This is AI-generated analysis, not legal advice."
            - "This opinion does not constitute a binding judgment or arbitration award."
            - "Consult a qualified attorney for legal advice specific to your situation."
            - "Analysis is based on the information provided and may not reflect all relevant facts or legal nuances."
    Payment:
      type: object
      properties:
        id:
          type: string
          pattern: ^pay_[a-z0-9]{3}$
          example: pay_bbb
        disputeId:
          type: string
        userId:
          type: string
        amountUsd:
          type: number
          format: decimal
          example: 99.00
        currency:
          type: string
          default: USD
        processor:
          type: string
          default: stripe
        processorPaymentId:
          type: string
        status:
          type: string
          enum: [pending, succeeded, failed, refunded, partially_refunded]
        refundedAmountUsd:
          type: number
          format: decimal
          default: 0
        refundReason:
          type: string
          nullable: true
        refundedAt:
          type: string
          format: date-time
          nullable: true
        idempotencyKey:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
          example: pi_xyz_secret_abc...
    ConfirmPaymentResponse:
      type: object
      properties:
        dispute:
          $ref: '#/components/schemas/Dispute'
    PaymentHistoryResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
    OpinionStatus:
      type: object
      properties:
        status:
          type: string
          enum: [in_progress, completed, failed]
        progress:
          type: object
          properties:
            evaluatorsComplete:
              type: integer
              example: 3
            evaluatorsTotal:
              type: integer
              example: 5
            aggregationStatus:
              type: string
              enum: [pending, in_progress, completed]
          required:
            - evaluatorsComplete
            - evaluatorsTotal
            - aggregationStatus
        estimatedCompletion:
          type: string
          format: date-time
          nullable: true
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
              requestId: req_xyz789
              documentationUrl: https://docs.meritview.app/errors/unauthorized
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: Email verification required
              requestId: req_xyz789
              documentationUrl: https://docs.meritview.app/errors/forbidden
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: The requested dispute does not exist or you do not have access
              details:
                dispute_id: disp_abc123
              requestId: req_xyz789
              documentationUrl: https://docs.meritview.app/errors/not-found
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: CONFLICT
              message: Dispute already has an active payment intent
              requestId: req_xyz789
              documentationUrl: https://docs.meritview.app/errors/conflict
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                email:
                  - Invalid email format
                password:
                  - Password must be at least 8 characters
              requestId: req_xyz789
              documentationUrl: https://docs.meritview.app/errors/validation-error
...
paths:
...
  /disputes:
    post:
      summary: Create a new dispute
      operationId: createDispute
      tags:
        - Disputes
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisputeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDisputeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'

    get:
      summary: List user's disputes
      operationId: listDisputes
      tags:
        - Disputes
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/DisputeState'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/PartyRole'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/DisputeCategory'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /disputes/{disputeId}:
    get:
      summary: Get dispute details
      operationId: getDispute
      tags:
        - Disputes
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
            pattern: ^disp_[a-z0-9]{6}$
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update dispute (draft state only)
      operationId: updateDispute
      tags:
        - Disputes
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDisputeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

    post:
      summary: Withdraw dispute
      operationId: withdrawDispute
      tags:
        - Disputes
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /disputes/{disputeId}/payment-intent:
    get:
      summary: Get Stripe payment intent
      operationId: getPaymentIntent
      tags:
        - Payments
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /disputes/{disputeId}/payment/confirm:
    post:
      summary: Confirm payment
      operationId: confirmPayment
      tags:
        - Payments
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmPaymentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /disputes/{disputeId}/refund-request:
    post:
      summary: Request refund
      operationId: requestRefund
      tags:
        - Payments
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refund request submitted for review
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/me/payments:
    get:
      summary: Get payment history
      operationId: getPaymentHistory
      tags:
        - Payments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentHistoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /disputes/{disputeId}/opinion:
    get:
      summary: Get opinion for completed dispute
      operationId: getOpinion
      tags:
        - Opinions
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opinion'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /disputes/{disputeId}/opinion/status:
    get:
      summary: Get opinion generation status
      operationId: getOpinionStatus
      tags:
        - Opinions
      security:
        - BearerAuth: []
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpinionStatus'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /webhooks/stripe:
    post:
      summary: Stripe webhook endpoint
      operationId: stripeWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    example: true
  /admin/disputes:
    get:
      summary: List all disputes (admin)
      operationId: adminListDisputes
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/DisputeState'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/DisputeCategory'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DisputeListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Dispute'
                            - type: object
                              properties:
                                _count:
                                  type: object
                                  properties:
                                    evaluatorOutputs:
                                      type: integer
                                    briefs:
                                      type: integer
                                    payments:
                                      type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/disputes/{id}:
    get:
      summary: Get dispute detail (admin)
      operationId: adminGetDisputeDetail
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dispute:
                    allOf:
                      - $ref: '#/components/schemas/Dispute'
                      - type: object
                        properties:
                          briefs:
                            type: array
                            items:
                              $ref: '#/components/schemas/Brief'
                          opinion:
                            $ref: '#/components/schemas/Opinion'
                          evaluatorOutputs:
                            type: array
                            items:
                              $ref: '#/components/schemas/EvaluatorOutput'
                          payments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...